requirements.txt
google-genai>=1.14.0
pydantic>=2.9.0
networkx>=3.4
rich>=13.8.0
nest-asyncio>=1.6.0
tenacity>=9.0.0
openai>=1.51.0
typing-extensions>=4.12.0
aiohttp>=3.8.0
